changes as per feedback
