+++ src/services/gasless.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/extensions */
-import { constants } from 'ethers';
+import { constants, BigNumber } from 'ethers';
 
 import { Token } from './tokensData';
 import { isPolygonAssetNative } from '../utils/blockchain';
@@ -47,5 +47,36 @@ export const getAllGaslessPaymasters = async (
   } catch (err) {
     console.error(err);
     return null;
+  }
+};
+
+export const getGasPrice = async (chainId: number) => {
+  let gasPrice = '0';
+  try {
+    const gasRes = await fetch(
+      `${process.env.REACT_APP_GAS_URL}/${chainId}?api-key=${process.env.REACT_APP_ETHERSPOT_BUNDLER_API_KEY}`,
+      {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          jsonrpc: '2.0',
+          id: 1,
+          method: 'skandha_getGasPrice',
+        }),
+      }
+    );
+    gasRes.json().then((response) => {
+      if (response.result) {
+        gasPrice = BigNumber.from(response.result.maxFeePerGas)
+          .add(response.result.maxPriorityFeePerGas)
+          .toString();
+      }
+    });
+    return gasPrice;
+  } catch (err) {
+    console.error(err);
+    return gasPrice;
   }
 };
